all: update tests and add integration test
The integration test will check that the auth flow works and that
it is checking the data base. it is a working example of a db
integration test as well as chceking that auth works. There were lots
of updates that needed to happen for the tests to work in both with
and without auth. some of the helpers are for testing and should be
kept out of the build.

Signed-off-by: soypete <petersonmiriah@gmail.com>